requirements.txt
firecrawl-py
pydantic
python-dotenv
